Changed type of return value from dot prod. No reason to use double precision there.

+++ src/embree.cpp
@@ -10,7 +10,7 @@ void rtc::init()
 double dot_prod( RTCRay &ray )
 {
 
-    double result = ray.dir[0]*ray.Ng[0];
+  float result = ray.dir[0]*ray.Ng[0];
   result += ray.dir[1]*ray.Ng[1];
   result += ray.dir[2]*ray.Ng[2];
